Refactor Branding related code
