+++ branding/collection/index.js
@@ -1,52 +1,3 @@
-Branding = new Mongo.Collection('branding');
+const Branding = new Mongo.Collection('branding');
 
-Schemas.BrandingSchema = new SimpleSchema({
-  projectLogoFileId: {
-    type: String,
-    optional: true
-  },
-  siteTitle: {
-    type: String,
-    optional: true
-  },
-  siteSlogan: {
-    type: String,
-    optional: true
-  },
-  siteFooter: {
-    type: String,
-    optional: true
-  },
-  socialMedia: {
-    type: [Object],
-    optional: true
-  },
-  "socialMedia.$.name": {
-    type: String,
-    allowedValues: ["Facebook", "Twitter", "Github"]
-  },
-  "socialMedia.$.url": {
-    type: String,
-    regEx: SimpleSchema.RegEx.Url,
-    autoform: {
-      placeholder: 'http://example.com/accountname'
-    }
-  }
-});
-
-// i18n translation
-Schemas.BrandingSchema.i18n("schemas.branding");
-
-Branding.attachSchema(Schemas.BrandingSchema);
-
-Branding.allow({
-  insert: function (userId) {
-    return Roles.userIsInRole(userId, ['admin']) && Branding.find().count() === 0;
-  },
-  update: function (userId) {
-    return Roles.userIsInRole(userId, ['admin']);
-  },
-  remove: function (userId) {
-    return Roles.userIsInRole(userId, ['admin']);
-  }
-});
+export { Branding };
