+++ branding/collection/permissions.js
@@ -1,0 +1,13 @@
+import { Branding } from './';
+
+Branding.allow({
+  insert: function (userId) {
+    return Roles.userIsInRole(userId, ['admin']) && Branding.find().count() === 0;
+  },
+  update: function (userId) {
+    return Roles.userIsInRole(userId, ['admin']);
+  },
+  remove: function (userId) {
+    return Roles.userIsInRole(userId, ['admin']);
+  }
+});
