+++ branding/server/publications.js
@@ -1,3 +1,5 @@
+import { Branding } from '/branding/collection';
+
 Meteor.publish('branding', function() {
   // Get Branding collection object
   return Branding.find({});
