+++ home/client/header/homeHeader.js
@@ -1,3 +1,4 @@
+import { Branding } from '/branding/collection';
 import { ProjectLogo } from '/branding/logo/collection';
 
 Template.homeHeader.onCreated(function() {
